// packages/react-native-sdk/src/theme/tokens/StreamTokens.types.ts
export type RNShadowToken = Required<
  Pick<
    ViewStyle,
    | 'elevation'
    | 'shadowColor'
    | 'shadowOffset'
    | 'shadowOpacity'
    | 'shadowRadius'
  >
>;

export interface FoundationsBlur {
  strong: number;
  subtle: number;
}

export interface FoundationsColors {
  baseBlack: ColorValue;
  baseTransparent0: ColorValue;
  baseTransparentBlack0: ColorValue;
  baseTransparentBlack10: ColorValue;
  baseTransparentBlack15: ColorValue;
  baseTransparentBlack20: ColorValue;
  baseTransparentBlack25: ColorValue;
  baseTransparentBlack5: ColorValue;
  baseTransparentBlack50: ColorValue;
  baseTransparentBlack70: ColorValue;
  baseTransparentBlack75: ColorValue;
  baseTransparentWhite0: ColorValue;
  baseTransparentWhite10: ColorValue;
  baseTransparentWhite15: ColorValue;
  baseTransparentWhite20: ColorValue;
  baseTransparentWhite25: ColorValue;
  baseTransparentWhite30: ColorValue;
  baseTransparentWhite50: ColorValue;
  baseTransparentWhite70: ColorValue;
  baseTransparentWhite75: ColorValue;
  baseWhite: ColorValue;
  blue100: ColorValue;
  blue150: ColorValue;
  blue200: ColorValue;
  blue300: ColorValue;
  blue400: ColorValue;
  blue50: ColorValue;
  blue500: ColorValue;
  blue600: ColorValue;
  blue700: ColorValue;
  blue800: ColorValue;
  blue900: ColorValue;
  cyan100: ColorValue;
  cyan150: ColorValue;
  cyan200: ColorValue;
  cyan300: ColorValue;
  cyan400: ColorValue;
  cyan50: ColorValue;
  cyan500: ColorValue;
  cyan600: ColorValue;
  cyan700: ColorValue;
  cyan800: ColorValue;
  cyan900: ColorValue;
  green100: ColorValue;
  green150: ColorValue;
  green200: ColorValue;
  green300: ColorValue;
  green400: ColorValue;
  green50: ColorValue;
  green500: ColorValue;
  green600: ColorValue;
  green700: ColorValue;
  green800: ColorValue;
  green900: ColorValue;
  lime100: ColorValue;
  lime150: ColorValue;
  lime200: ColorValue;
  lime300: ColorValue;
  lime400: ColorValue;
  lime50: ColorValue;
  lime500: ColorValue;
  lime600: ColorValue;
  lime700: ColorValue;
  lime800: ColorValue;
  lime900: ColorValue;
  neutral100: ColorValue;
  neutral150: ColorValue;
  neutral200: ColorValue;
  neutral300: ColorValue;
  neutral400: ColorValue;
  neutral50: ColorValue;
  neutral500: ColorValue;
  neutral600: ColorValue;
  neutral700: ColorValue;
  neutral800: ColorValue;
  neutral900: ColorValue;
  purple100: ColorValue;
  purple150: ColorValue;
  purple200: ColorValue;
  purple300: ColorValue;
  purple400: ColorValue;
  purple50: ColorValue;
  purple500: ColorValue;
  purple600: ColorValue;
  purple700: ColorValue;
  purple800: ColorValue;
  purple900: ColorValue;
  red100: ColorValue;
  red150: ColorValue;
  red200: ColorValue;
  red300: ColorValue;
  red400: ColorValue;
  red50: ColorValue;
  red500: ColorValue;
  red600: ColorValue;
  red700: ColorValue;
  red800: ColorValue;
  red900: ColorValue;
  slate100: ColorValue;
  slate150: ColorValue;
  slate200: ColorValue;
  slate300: ColorValue;
  slate400: ColorValue;
  slate50: ColorValue;
  slate500: ColorValue;
  slate600: ColorValue;
  slate700: ColorValue;
  slate800: ColorValue;
  slate900: ColorValue;
  violet100: ColorValue;
  violet150: ColorValue;
  violet200: ColorValue;
  violet300: ColorValue;
  violet400: ColorValue;
  violet50: ColorValue;
  violet500: ColorValue;
  violet600: ColorValue;
  violet700: ColorValue;
  violet800: ColorValue;
  violet900: ColorValue;
  yellow100: ColorValue;
  yellow150: ColorValue;
  yellow200: ColorValue;
  yellow300: ColorValue;
  yellow400: ColorValue;
  yellow50: ColorValue;
  yellow500: ColorValue;
  yellow600: ColorValue;
  yellow700: ColorValue;
  yellow800: ColorValue;
  yellow900: ColorValue;
}

export interface FoundationsLayout {
  size10: number;
  size12: number;
  size128: number;
  size13: number;
  size14: number;
  size144: number;
  size15: number;
  size16: number;
  size17: number;
  size18: number;
  size2: number;
  size20: number;
  size208: number;
  size22: number;
  size24: number;
  size240: number;
  size28: number;
  size32: number;
  size320: number;
  size4: number;
  size40: number;
  size48: number;
  size480: number;
  size56: number;
  size560: number;
  size6: number;
  size64: number;
  size640: number;
  size760: number;
  size8: number;
  size80: number;
}

export interface FoundationsRadius {
  radius0: number;
  radius12: number;
  radius16: number;
  radius2: number;
  radius20: number;
  radius24: number;
  radius32: number;
  radius4: number;
  radius6: number;
  radius8: number;
  radiusFull: number;
}

export interface FoundationsSpacing {
  space0: number;
  space10: number;
  space12: number;
  space14: number;
  space16: number;
  space2: number;
  space20: number;
  space24: number;
  space32: number;
  space4: number;
  space40: number;
  space48: number;
  space6: number;
  space64: number;
  space8: number;
  space80: number;
}

export interface FoundationsStroke {
  w100: number;
  w120: number;
  w150: number;
  w200: number;
  w300: number;
  w400: number;
  w80: number;
}

export interface FoundationsTypography {
  fontFamilyDmMono: TextStyle['fontFamily'];
  fontFamilyGeist: TextStyle['fontFamily'];
  fontFamilyGeistMono: TextStyle['fontFamily'];
  fontFamilyInter: TextStyle['fontFamily'];
  fontFamilyRoboto: TextStyle['fontFamily'];
  fontFamilyRobotoMono: TextStyle['fontFamily'];
  fontFamilySfMono: TextStyle['fontFamily'];
  fontFamilySfPro: TextStyle['fontFamily'];
  fontSizeSize10: NonNullable<TextStyle['fontSize']>;
  fontSizeSize11: NonNullable<TextStyle['fontSize']>;
  fontSizeSize12: NonNullable<TextStyle['fontSize']>;
  fontSizeSize13: NonNullable<TextStyle['fontSize']>;
  fontSizeSize14: NonNullable<TextStyle['fontSize']>;
  fontSizeSize15: NonNullable<TextStyle['fontSize']>;
  fontSizeSize16: NonNullable<TextStyle['fontSize']>;
  fontSizeSize17: NonNullable<TextStyle['fontSize']>;
  fontSizeSize18: NonNullable<TextStyle['fontSize']>;
  fontSizeSize20: NonNullable<TextStyle['fontSize']>;
  fontSizeSize22: NonNullable<TextStyle['fontSize']>;
  fontSizeSize24: NonNullable<TextStyle['fontSize']>;
  fontSizeSize28: NonNullable<TextStyle['fontSize']>;
  fontSizeSize32: NonNullable<TextStyle['fontSize']>;
  fontSizeSize40: NonNullable<TextStyle['fontSize']>;
  fontSizeSize48: NonNullable<TextStyle['fontSize']>;
  fontSizeSize64: NonNullable<TextStyle['fontSize']>;
  fontSizeSize8: NonNullable<TextStyle['fontSize']>;
  fontWeightW400: TextStyle['fontWeight'];
  fontWeightW500: TextStyle['fontWeight'];
  fontWeightW600: TextStyle['fontWeight'];
  fontWeightW700: TextStyle['fontWeight'];
  lineHeightLineHeight10: TextStyle['lineHeight'];
  lineHeightLineHeight12: TextStyle['lineHeight'];
  lineHeightLineHeight13: TextStyle['lineHeight'];
  lineHeightLineHeight14: TextStyle['lineHeight'];
  lineHeightLineHeight15: TextStyle['lineHeight'];
  lineHeightLineHeight16: TextStyle['lineHeight'];
  lineHeightLineHeight17: TextStyle['lineHeight'];
  lineHeightLineHeight18: TextStyle['lineHeight'];
  lineHeightLineHeight20: TextStyle['lineHeight'];
  lineHeightLineHeight24: TextStyle['lineHeight'];
  lineHeightLineHeight28: TextStyle['lineHeight'];
  lineHeightLineHeight32: TextStyle['lineHeight'];
  lineHeightLineHeight40: TextStyle['lineHeight'];
  lineHeightLineHeight48: TextStyle['lineHeight'];
  lineHeightLineHeight8: TextStyle['lineHeight'];
}

export interface ChatComponents {
  buttonHitTargetMinHeight: number;
  buttonHitTargetMinWidth: number;
  buttonPaddingXIconOnlyLg: number;
  buttonPaddingXIconOnlyMd: number;
  buttonPaddingXIconOnlySm: number;
  buttonPaddingXIconOnlyXs: number;
  buttonPaddingXWithLabelLg: number;
  buttonPaddingXWithLabelMd: number;
  buttonPaddingXWithLabelSm: number;
  buttonPaddingXWithLabelXs: number;
  buttonPaddingYLg: number;
  buttonPaddingYMd: number;
  buttonPaddingYSm: number;
  buttonPaddingYXs: number;
  buttonRadiusFull: number;
  buttonRadiusLg: number;
  buttonRadiusMd: number;
  buttonRadiusSm: number;
  buttonVisualHeightLg: number;
  buttonVisualHeightMd: number;
  buttonVisualHeightSm: number;
  buttonVisualHeightXs: number;
  composerRadiusFixed: number;
  composerRadiusFloating: number;
  deviceRadius: number;
  deviceSafeAreaBottom: number;
  deviceSafeAreaTop: number;
  emoji2xl: number;
  emojiLg: number;
  emojiMd: number;
  emojiSm: number;
  emojiXl: number;
  iconSizeLg: number;
  iconSizeMd: number;
  iconSizeSm: number;
  iconSizeXs: number;
  iconStrokeDefault: number;
  iconStrokeEmphasis: number;
  iconStrokeSubtle: number;
  inputRadiusOptionCard: number;
  inputRadiusPollOptionInput: number;
  inputRadiusSearchInput: number;
  inputRadiusSelectInput: number;
  inputRadiusTextInput: number;
  messageBubbleRadiusAttachment: number;
  messageBubbleRadiusAttachmentInline: number;
  messageBubbleRadiusGroupBottom: number;
  messageBubbleRadiusGroupMiddle: number;
  messageBubbleRadiusGroupTop: number;
  messageBubbleRadiusTail: number;
}

export interface ChatPrimitives {
  darkElevation1: RNShadowToken;
  darkElevation2: RNShadowToken;
  darkElevation3: RNShadowToken;
  darkElevation4: RNShadowToken;
  lightElevation1: RNShadowToken;
  lightElevation2: RNShadowToken;
  lightElevation3: RNShadowToken;
  lightElevation4: RNShadowToken;
  radius2xl: number;
  radius3xl: number;
  radius4xl: number;
  radiusLg: number;
  radiusMax: number;
  radiusMd: number;
  radiusNone: number;
  radiusSm: number;
  radiusXl: number;
  radiusXs: number;
  radiusXxs: number;
  spacing2xl: number;
  spacing3xl: number;
  spacingLg: number;
  spacingMd: number;
  spacingNone: number;
  spacingSm: number;
  spacingXl: number;
  spacingXs: number;
  spacingXxs: number;
  spacingXxxs: number;
  typographyFontFamilyMono: TextStyle['fontFamily'];
  typographyFontFamilySans: TextStyle['fontFamily'];
  typographyFontSize2xl: NonNullable<TextStyle['fontSize']>;
  typographyFontSizeLg: NonNullable<TextStyle['fontSize']>;
  typographyFontSizeMd: NonNullable<TextStyle['fontSize']>;
  typographyFontSizeMicro: NonNullable<TextStyle['fontSize']>;
  typographyFontSizeSm: NonNullable<TextStyle['fontSize']>;
  typographyFontSizeXl: NonNullable<TextStyle['fontSize']>;
  typographyFontSizeXs: NonNullable<TextStyle['fontSize']>;
  typographyFontSizeXxs: NonNullable<TextStyle['fontSize']>;
  typographyFontWeightBold: TextStyle['fontWeight'];
  typographyFontWeightMedium: TextStyle['fontWeight'];
  typographyFontWeightRegular: TextStyle['fontWeight'];
  typographyFontWeightSemiBold: TextStyle['fontWeight'];
  typographyLineHeightNormal: TextStyle['lineHeight'];
  typographyLineHeightRelaxed: TextStyle['lineHeight'];
  typographyLineHeightTight: TextStyle['lineHeight'];
}
// ...
export interface IStreamTokens {
  foundations: {
    blur: FoundationsBlur;
    colors: FoundationsColors;
    layout: FoundationsLayout;
    radius: FoundationsRadius;
    spacing: FoundationsSpacing;
    stroke: FoundationsStroke;
    typography: FoundationsTypography;
  };
  components: ChatComponents;
  primitives: ChatPrimitives;
  semantics: ChatSemantics;
}